[nzbget] Allow mounting additional volumes
This is useful if some data exists in different paths and you wish to make use
of categories + paths. So you can have a volume for Movies and have a movies
category that points to this path.

Useful for when they exist on NFS, this allows you to easily attach NFS mounts
to the pod.

Signed-off-by: Yasser Saleemi <yassersaleemi@gmail.com>